iproute2: lib: fix crash seen while running tc q_htb with rate specified
While running tc q_htb with rate specified, a crash is seen.
signal 11 (SIGSEGV), code 2 (SEGV_ACCERR)
The exact location of the crash is get_hz() where the extern int
is assigned the value from the function __get_hz().
The current suspicion is that the issue could be because of the
declaration of the variable util.h and for some reason the definition
in util.c is not being read. Explictly assigning a value to this
variable seems to be allocating the memory and no crash is seen.

CRs-fixed: 665103
Change-Id: Ia6c1d7f33978f7831b662109eb211fe771c59a14